Remove WPCOM overview components and move utils and hooks to shared folder.
