+++ client/a8c-for-agencies/sections/marketplace/lib/wpcom-bulk-values-utils.ts
@@ -1,4 +1,8 @@
-import { Option } from '../wpcom-slider';
+type Option = {
+	label: string;
+	value: number | string | null;
+	sub?: string;
+};
 
 type SliderOption = {
 	minAbsolute: number;
