+++ client/a8c-for-agencies/sections/marketplace/shopping-cart/shopping-cart-menu/item.tsx
@@ -4,7 +4,7 @@ import { Icon, check } from '@wordpress/icons';
 import { useTranslate } from 'i18n-calypso';
 import { useSelector } from 'calypso/state';
 import { getProductsList } from 'calypso/state/products-list/selectors';
-import { useGetProductPricingInfo } from '../../wpcom-overview/hooks/use-total-invoice-value';
+import { useGetProductPricingInfo } from '../../hooks/use-total-invoice-value';
 import type { ShoppingCartItem } from '../../types';
 
 import './style.scss';
