--- client/a8c-for-agencies/sections/marketplace/wpcom-overview/style.scss
@@ -1,50 +1,0 @@
-@import "@wordpress/base-styles/breakpoints";
-@import "@wordpress/base-styles/mixins";
-
-.wpcom-overview .hosting-dashboard-layout__body-wrapper {
-	display: flex;
-	flex-direction: column;
-	gap: 64px;
-	margin-block-start: 48px;
-	margin-block-end: 64px;
-
-	.bulk-selection {
-		width: 100%;
-
-		@include break-large {
-			width: 800px;
-			margin-inline: auto;
-		}
-	}
-}
-
-.hosting-dashboard-layout__header-breadcrumb {
-	@include breakpoint-deprecated( "<660px" ) {
-		display: none;
-	}
-}
-
-.bulk-selection {
-	text-align: center;
-}
-
-.bulk-selection__owned-plan {
-	display: inline-flex;
-	flex-direction: row;
-	gap: 16px;
-
-	background-color: var(--color-neutral-0);
-	padding: 8px 16px;
-	margin: 0 auto 32px;
-}
-
-
-.bulk-selection.is-placeholder .bulk-selection__slider {
-	animation: loading-fade 1.6s ease-in-out infinite;
-	background: var(--color-neutral-10);
-	max-width: 800px;
-	width: 100%;
-	height: 80.8px;
-	margin: 0 auto;
-	border-radius: 4px;
-}
