--- client/a8c-for-agencies/sections/marketplace/wpcom-overview/wpcom-card/style.scss
@@ -1,103 +1,0 @@
-@import "@wordpress/base-styles/breakpoints";
-@import "@wordpress/base-styles/mixins";
-
-.wpcom-plan-card {
-	border: 1px solid var(--color-neutral-5);
-	border-radius: 4px;
-	margin: 0;
-	padding: 0;
-}
-
-.wpcom-plan-card__section {
-	padding: 48px;
-	display: flex;
-	flex-direction: column;
-	gap: 32px;
-	align-items: flex-start;
-
-	&.is-main {
-		background: var(--color-neutral-0);
-	}
-
-	&:last-child {
-		border-block-start: 1px solid var(--color-neutral-5);
-	}
-}
-
-.wpcom-plan-card__title {
-	font-size: 1.125rem; // stylelint-disable-line
-	font-weight: 510; // stylelint-disable-line
-	line-height: 1.5;
-	display: flex;
-	align-items: center;
-	gap: 10px;
-
-	.jetpack-logo {
-		display: none;
-
-		@include break-large {
-			display: block;
-		}
-	}
-
-}
-
-.wpcom-plan-card__price-actual-value {
-	color: var(--color-black);
-	font-size: 2rem;
-	font-weight: 600;
-	line-height: 1;
-}
-
-.wpcom-plan-card__price-original-value {
-	display: inline-block;
-	color: var(--color-neutral-80);
-	font-size: 1.5rem;
-	font-weight: 400;
-	line-height: 1;
-	text-decoration: line-through;
-	margin-inline: 8px;
-}
-
-.wpcom-plan-card__price-discount {
-	color: var(--color-success);
-}
-
-.wpcom-plan-card__price-interval {
-	font-size: 0.75rem;
-	line-height: 1.1;
-	font-weight: 400;
-}
-
-.wpcom-plan-card__features {
-	width: 100%;
-	display: grid;
-	grid-template-columns: 1fr;
-	gap: 4px;
-
-	@include break-large {
-		grid-template-columns: 1fr 1fr;
-		gap: 48px;
-	}
-}
-
-.wpcom-plan-card__features.is-jetpack .simple-list-icon {
-	fill: var(--color-jetpack);
-}
-
-.wpcom-plan-card__cta.is-placeholder,
-.wpcom-plan-card__price.is-placeholder {
-	animation: loading-fade 1.6s ease-in-out infinite;
-	background: var(--color-neutral-10);
-	border-radius: 4px;
-}
-
-.wpcom-plan-card__cta.is-placeholder {
-	width: 208px;
-	height: 40px;
-}
-
-.wpcom-plan-card__price.is-placeholder {
-	width: 154px;
-	height: 47.2px;
-}
